databroker: cleanup
